==> dos/03-dta-icd10-113-cause-list.R <==

library(tidyverse)

"
GR113-000  |  All causes
GR113-001  |  #Salmonella infections (A01-A02)  
GR113-002  |  #Shigellosis and amebiasis (A03,A06)  
GR113-003  |  Certain other intestinal infections (A04,A07-A09)  
GR113-004  |  #Tuberculosis (A16-A19)  
GR113-005  |     Respiratory tuberculosis (A16)  
GR113-006  |     Other tuberculosis (A17-A19)  
GR113-007  |  #Whooping cough (A37)  
GR113-008  |  #Scarlet fever and erysipelas (A38,A46)  
GR113-009  |  #Meningococcal infection (A39)  
GR113-010  |  #Septicemia (A40-A41)  
GR113-011  |  #Syphilis (A50-A53)  
GR113-012  |  #Acute poliomyelitis (A80)  
GR113-013  |  #Arthropod-borne viral encephalitis (A83-A84,A85.2)  
GR113-014  |  #Measles (B05)  
GR113-015  |  #Viral hepatitis (B15-B19)  
GR113-016  |  #Human immunodeficiency virus (HIV) disease (B20-B24)  
GR113-017  |  #Malaria (B50-B54)  
GR113-018  |  Other and unspecified infectious and parasitic diseases (A00...)  
GR113-019  |  #Malignant neoplasms (C00-C97)  
GR113-020  |     Malignant neoplasms of lip, oral cavity and pharynx (C00-C14)  
GR113-021  |     Malignant neoplasm of esophagus (C15)  
GR113-022  |     Malignant neoplasm of stomach (C16)  
GR113-023  |     Malignant neoplasms of colon, rectum and anus (C18-C21)  
GR113-024  |     Malignant neoplasms of liver and intrahepatic bile ducts (C22)  
GR113-025  |     Malignant neoplasm of pancreas (C25)  
GR113-026  |     Malignant neoplasm of larynx (C32)  
GR113-027  |     Malignant neoplasms of trachea, bronchus and lung (C33-C34)  
GR113-028  |     Malignant melanoma of skin (C43)  
GR113-029  |     Malignant neoplasm of breast (C50)  
GR113-030  |     Malignant neoplasm of cervix uteri (C53)  
GR113-031  |     Malignant neoplasms of corpus uteri and uterus, part unspecified (C54-C55)  
GR113-032  |     Malignant neoplasm of ovary (C56)  
GR113-033  |     Malignant neoplasm of prostate (C61)  
GR113-034  |     Malignant neoplasms of kidney and renal pelvis (C64-C65)  
GR113-035  |     Malignant neoplasm of bladder (C67)  
GR113-036  |     Malignant neoplasms of meninges, brain and other parts of central nervous system (C70-C72)  
GR113-037  |     Malignant neoplasms of lymphoid, hematopoietic and related tissue (C81-C96)  
GR113-038  |        Hodgkin's disease (C81)  
GR113-039  |        Non-Hodgkin's lymphoma (C82-C85)  
GR113-040  |        Leukemia (C91-C95)  
GR113-041  |        Multiple myeloma and immunoproliferative neoplasms (C88,C90)  
GR113-042  |        Other and unspecified malignant neoplasms of lymphoid, hematopoietic and related tissue (C96)  
GR113-043  |     All other and unspecified malignant neoplasms (C17...)  
GR113-044  |  #In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior (D00-D48)  
GR113-045  |  #Anemias (D50-D64)  
GR113-046  |  #Diabetes mellitus (E10-E14)  
GR113-047  |  #Nutritional deficiencies (E40-E64)  
GR113-048  |     Malnutrition (E40-E46)  
GR113-049  |     Other nutritional deficiencies (E50-E64)  
GR113-050  |  #Meningitis (G00,G03)  
GR113-051  |  #Parkinson's disease (G20-G21)  
GR113-052  |  #Alzheimer's disease (G30)  
GR113-053  |  Major cardiovascular diseases (I00-I78)  
GR113-054  |     #Diseases of heart (I00-I09,I11,I13,I20-I51)  
GR113-055  |        Acute rheumatic fever and chronic rheumatic heart diseases (I00-I09)  
GR113-056  |        Hypertensive heart disease (I11)  
GR113-057  |        Hypertensive heart and renal disease (I13)  
GR113-058  |        Ischemic heart diseases (I20-I25)  
GR113-059  |           Acute myocardial infarction (I21-I22)  
GR113-060  |           Other acute ischemic heart diseases (I24)  
GR113-061  |           Other forms of chronic ischemic heart disease (I20,I25)  
GR113-062  |              Atherosclerotic cardiovascular disease, so described (I25.0)  
GR113-063  |              All other forms of chronic ischemic heart disease (I20,I25.1-I25.9)  
GR113-064  |        Other heart diseases (I26-I51)  
GR113-065  |           Acute and subacute endocarditis (I33)  
GR113-066  |           Diseases of pericardium and acute myocarditis (I30-I31,I40)  
GR113-067  |           Heart failure (I50)  
GR113-068  |           All other forms of heart disease (I26-I28,I34-I38,I42-I49,I51)  
GR113-069  |     #Essential hypertension and hypertensive renal disease (I10,I12,I15)  
GR113-070  |     #Cerebrovascular diseases (I60-I69)  
GR113-071  |     #Atherosclerosis (I70)  
GR113-072  |     Other diseases of circulatory system (I71-I78)  
GR113-073  |        #Aortic aneurysm and dissection (I71)  
GR113-074  |        Other diseases of arteries, arterioles and capillaries (I72-I78)  
GR113-075  |  Other disorders of circulatory system (I80-I99)  
GR113-076  |  #Influenza and pneumonia (J09-J18)  
GR113-077  |     Influenza (J09-J11)  
GR113-078  |     Pneumonia (J12-J18)  
GR113-079  |  Other acute lower respiratory infections (J20-J22,U04)  
GR113-080  |     #Acute bronchitis and bronchiolitis (J20-J21)  
GR113-081  |     Other and unspecified acute lower respiratory infections (J22,U04)  
GR113-082  |  #Chronic lower respiratory diseases (J40-J47)  
GR113-083  |     Bronchitis, chronic and unspecified (J40-J42)  
GR113-084  |     Emphysema (J43)  
GR113-085  |     Asthma (J45-J46)  
GR113-086  |     Other chronic lower respiratory diseases (J44,J47)  
GR113-087  |  #Pneumoconioses and chemical effects (J60-J66,J68)  
GR113-088  |  #Pneumonitis due to solids and liquids (J69)  
GR113-089  |  Other diseases of respiratory system (J00-J06,J30- J39,J67,J70-J98)  
GR113-090  |  #Peptic ulcer (K25-K28)  
GR113-091  |  #Diseases of appendix (K35-K38)  
GR113-092  |  #Hernia (K40-K46)  
GR113-093  |  #Chronic liver disease and cirrhosis (K70,K73-K74)  
GR113-094  |     Alcoholic liver disease (K70)  
GR113-095  |     Other chronic liver disease and cirrhosis (K73-K74)  
GR113-096  |  #Cholelithiasis and other disorders of gallbladder (K80-K82)  
GR113-097  |  #Nephritis, nephrotic syndrome and nephrosis (N00-N07,N17-N19,N25-N27)  
GR113-098  |     Acute and rapidly progressive nephritic and nephrotic syndrome (N00-N01,N04)  
GR113-099  |     Chronic glomerulonephritis, nephritis and nephropathy not specified as acute or chronic, and renal sclerosis unspecified (N02-N03,N05-N07,N26)  
GR113-100  |     Renal failure (N17-N19)  
GR113-101  |     Other disorders of kidney (N25,N27)  
GR113-102  |  #Infections of kidney (N10-N12,N13.6,N15.1)  
GR113-103  |  #Hyperplasia of prostate (N40)  
GR113-104  |  #Inflammatory diseases of female pelvic organs (N70-N76)  
GR113-105  |  #Pregnancy, childbirth and the puerperium (O00-O99)  
GR113-106  |     Pregnancy with abortive outcome (O00-O07)  
GR113-107  |     Other complications of pregnancy, childbirth and the puerperium (O10-O99)  
GR113-108  |  #Certain conditions originating in the perinatal period (P00-P96)  
GR113-109  |  #Congenital malformations, deformations and chromosomal abnormalities (Q00-Q99)  
GR113-110  |  Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified (R00-R99)  
GR113-111  |  All other diseases (Residual)   
GR113-112  |  #Accidents (unintentional injuries) (V01-X59,Y85-Y86)  
GR113-113  |     Transport accidents (V01-V99,Y85)  
GR113-114  |        Motor vehicle accidents (V02-V04...)  
GR113-115  |        Other land transport accidents (V01...)  
GR113-116  |        Water, air and space, and other and unspecified transport accidents and their sequelae (V90-V99,Y85)  
GR113-117  |     Nontransport accidents (W00-X59,Y86)  
GR113-118  |        Falls (W00-W19)  
GR113-119  |        Accidental discharge of firearms (W32-W34)  
GR113-120  |        Accidental drowning and submersion (W65-W74)  
GR113-121  |        Accidental exposure to smoke, fire and flames (X00-X09)  
GR113-122  |        Accidental poisoning and exposure to noxious substances (X40-X49)  
GR113-123  |        Other and unspecified nontransport accidents and their sequelae (W20-W31,W35-W64,W75-W99,X10-X39,X50-X59,Y86)  
GR113-124  |  #Intentional self-harm (suicide) (*U03,X60-X84,Y87.0)  
GR113-125  |     Intentional self-harm (suicide) by discharge of firearms (X72-X74)  
GR113-126  |     Intentional self-harm (suicide) by other and unspecified means and their sequelae (*U03,X60-X71,X75-X84,Y87.0)  
GR113-127  |  #Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)  
GR113-128  |     Assault (homicide) by discharge of firearms (*U01.4,X93-X95)  
GR113-129  |     Assault (homicide) by other and unspecified means and their sequelae (*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1)  
GR113-130  |  #Legal intervention (Y35,Y89.0)  
GR113-131  |  Events of undetermined intent (Y10-Y34,Y87.2,Y89.9)  
GR113-132  |     Discharge of firearms, undetermined intent (Y22-Y24)  
GR113-133  |     Other and unspecified events of undetermined intent and their sequelae (Y10-Y21,Y25-Y34,Y87.2,Y89.9)  
GR113-134  |  #Operations of war and their sequelae (Y36,Y89.1)  
GR113-135  |  #Complications of medical and surgical care (Y40-Y84,Y88)  
GR113-136  |  #Enterocolitis due to Clostridium difficile (A04.7)" %>% 
  read_lines() %>% 
  strsplit("|", fixed = TRUE) %>% 
  do.call(rbind, .) %>% 
  as_tibble() %>% 
  mutate_all(stringr::str_trim) %>% 
  mutate(V2 = tolower(V2)) %>% 
  filter(V1 %in% c(
    "GR113-000", # All
    "GR113-028", # Skin Cancer
    "GR113-046", # Diabetes Mellitus
    "GR113-054", # Diseases of Heart
    "GR113-058", # Ischemic Heart Disease
    "GR113-070", # Cerebrovascular Diseases
    "GR113-076", # Influenza and Pneumonia
    "GR113-124", # Suicide
    "GR113-112", # Accidents (unintentional injuries) (V01-X59,Y85-Y86)  
    "GR113-113", # Transport accidents (V01-V99,Y85)  
    "GR113-117", # Nontransport accidents (W00-X59,Y86)
    "GR113-127"  # Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)
  )) %>% 
  mutate(V3 = case_when(
    V1 == "GR113-000" ~ "all_causes",
    V1 == "GR113-028" ~ "skin_cancer",
    V1 == "GR113-046" ~ "diabetes_mellitus",
    V1 == "GR113-054" ~ "disease_of_heart",
    V1 == "GR113-058" ~ "ischemic_heart_disease",
    V1 == "GR113-070" ~ "cerebrovascular_diseases",
    V1 == "GR113-076" ~ "influenza_and_pneumonia",
    V1 == "GR113-124" ~ "suicide",
    V1 == "GR113-112" ~ "accidents",
    V1 == "GR113-113" ~ "transport_acc",
    V1 == "GR113-117" ~ "non-trans_acc",
    V1 == "GR113-127" ~ "assualt"
  )) %>% 
  print() %>% 
  write_csv("dta/03-icd10-113-cause-list.csv")

# Placebo causes of death -------------------------------------------------

"
GR113-000  |  All causes
GR113-001  |  #Salmonella infections (A01-A02)  
GR113-002  |  #Shigellosis and amebiasis (A03,A06)  
GR113-003  |  Certain other intestinal infections (A04,A07-A09)  
GR113-004  |  #Tuberculosis (A16-A19)  
GR113-005  |     Respiratory tuberculosis (A16)  
GR113-006  |     Other tuberculosis (A17-A19)  
GR113-007  |  #Whooping cough (A37)  
GR113-008  |  #Scarlet fever and erysipelas (A38,A46)  
GR113-009  |  #Meningococcal infection (A39)  
GR113-010  |  #Septicemia (A40-A41)  
GR113-011  |  #Syphilis (A50-A53)  
GR113-012  |  #Acute poliomyelitis (A80)  
GR113-013  |  #Arthropod-borne viral encephalitis (A83-A84,A85.2)  
GR113-014  |  #Measles (B05)  
GR113-015  |  #Viral hepatitis (B15-B19)  
GR113-016  |  #Human immunodeficiency virus (HIV) disease (B20-B24)  
GR113-017  |  #Malaria (B50-B54)  
GR113-018  |  Other and unspecified infectious and parasitic diseases (A00...)  
GR113-019  |  #Malignant neoplasms (C00-C97)  
GR113-020  |     Malignant neoplasms of lip, oral cavity and pharynx (C00-C14)  
GR113-021  |     Malignant neoplasm of esophagus (C15)  
GR113-022  |     Malignant neoplasm of stomach (C16)  
GR113-023  |     Malignant neoplasms of colon, rectum and anus (C18-C21)  
GR113-024  |     Malignant neoplasms of liver and intrahepatic bile ducts (C22)  
GR113-025  |     Malignant neoplasm of pancreas (C25)  
GR113-026  |     Malignant neoplasm of larynx (C32)  
GR113-027  |     Malignant neoplasms of trachea, bronchus and lung (C33-C34)  
GR113-028  |     Malignant melanoma of skin (C43)  
GR113-029  |     Malignant neoplasm of breast (C50)  
GR113-030  |     Malignant neoplasm of cervix uteri (C53)  
GR113-031  |     Malignant neoplasms of corpus uteri and uterus, part unspecified (C54-C55)  
GR113-032  |     Malignant neoplasm of ovary (C56)  
GR113-033  |     Malignant neoplasm of prostate (C61)  
GR113-034  |     Malignant neoplasms of kidney and renal pelvis (C64-C65)  
GR113-035  |     Malignant neoplasm of bladder (C67)  
GR113-036  |     Malignant neoplasms of meninges, brain and other parts of central nervous system (C70-C72)  
GR113-037  |     Malignant neoplasms of lymphoid, hematopoietic and related tissue (C81-C96)  
GR113-038  |        Hodgkin's disease (C81)  
GR113-039  |        Non-Hodgkin's lymphoma (C82-C85)  
GR113-040  |        Leukemia (C91-C95)  
GR113-041  |        Multiple myeloma and immunoproliferative neoplasms (C88,C90)  
GR113-042  |        Other and unspecified malignant neoplasms of lymphoid, hematopoietic and related tissue (C96)  
GR113-043  |     All other and unspecified malignant neoplasms (C17...)  
GR113-044  |  #In situ neoplasms, benign neoplasms and neoplasms of uncertain or unknown behavior (D00-D48)  
GR113-045  |  #Anemias (D50-D64)  
GR113-046  |  #Diabetes mellitus (E10-E14)  
GR113-047  |  #Nutritional deficiencies (E40-E64)  
GR113-048  |     Malnutrition (E40-E46)  
GR113-049  |     Other nutritional deficiencies (E50-E64)  
GR113-050  |  #Meningitis (G00,G03)  
GR113-051  |  #Parkinson's disease (G20-G21)  
GR113-052  |  #Alzheimer's disease (G30)  
GR113-053  |  Major cardiovascular diseases (I00-I78)  
GR113-054  |     #Diseases of heart (I00-I09,I11,I13,I20-I51)  
GR113-055  |        Acute rheumatic fever and chronic rheumatic heart diseases (I00-I09)  
GR113-056  |        Hypertensive heart disease (I11)  
GR113-057  |        Hypertensive heart and renal disease (I13)  
GR113-058  |        Ischemic heart diseases (I20-I25)  
GR113-059  |           Acute myocardial infarction (I21-I22)  
GR113-060  |           Other acute ischemic heart diseases (I24)  
GR113-061  |           Other forms of chronic ischemic heart disease (I20,I25)  
GR113-062  |              Atherosclerotic cardiovascular disease, so described (I25.0)  
GR113-063  |              All other forms of chronic ischemic heart disease (I20,I25.1-I25.9)  
GR113-064  |        Other heart diseases (I26-I51)  
GR113-065  |           Acute and subacute endocarditis (I33)  
GR113-066  |           Diseases of pericardium and acute myocarditis (I30-I31,I40)  
GR113-067  |           Heart failure (I50)  
GR113-068  |           All other forms of heart disease (I26-I28,I34-I38,I42-I49,I51)  
GR113-069  |     #Essential hypertension and hypertensive renal disease (I10,I12,I15)  
GR113-070  |     #Cerebrovascular diseases (I60-I69)  
GR113-071  |     #Atherosclerosis (I70)  
GR113-072  |     Other diseases of circulatory system (I71-I78)  
GR113-073  |        #Aortic aneurysm and dissection (I71)  
GR113-074  |        Other diseases of arteries, arterioles and capillaries (I72-I78)  
GR113-075  |  Other disorders of circulatory system (I80-I99)  
GR113-076  |  #Influenza and pneumonia (J09-J18)  
GR113-077  |     Influenza (J09-J11)  
GR113-078  |     Pneumonia (J12-J18)  
GR113-079  |  Other acute lower respiratory infections (J20-J22,U04)  
GR113-080  |     #Acute bronchitis and bronchiolitis (J20-J21)  
GR113-081  |     Other and unspecified acute lower respiratory infections (J22,U04)  
GR113-082  |  #Chronic lower respiratory diseases (J40-J47)  
GR113-083  |     Bronchitis, chronic and unspecified (J40-J42)  
GR113-084  |     Emphysema (J43)  
GR113-085  |     Asthma (J45-J46)  
GR113-086  |     Other chronic lower respiratory diseases (J44,J47)  
GR113-087  |  #Pneumoconioses and chemical effects (J60-J66,J68)  
GR113-088  |  #Pneumonitis due to solids and liquids (J69)  
GR113-089  |  Other diseases of respiratory system (J00-J06,J30- J39,J67,J70-J98)  
GR113-090  |  #Peptic ulcer (K25-K28)  
GR113-091  |  #Diseases of appendix (K35-K38)  
GR113-092  |  #Hernia (K40-K46)  
GR113-093  |  #Chronic liver disease and cirrhosis (K70,K73-K74)  
GR113-094  |     Alcoholic liver disease (K70)  
GR113-095  |     Other chronic liver disease and cirrhosis (K73-K74)  
GR113-096  |  #Cholelithiasis and other disorders of gallbladder (K80-K82)  
GR113-097  |  #Nephritis, nephrotic syndrome and nephrosis (N00-N07,N17-N19,N25-N27)  
GR113-098  |     Acute and rapidly progressive nephritic and nephrotic syndrome (N00-N01,N04)  
GR113-099  |     Chronic glomerulonephritis, nephritis and nephropathy not specified as acute or chronic, and renal sclerosis unspecified (N02-N03,N05-N07,N26)  
GR113-100  |     Renal failure (N17-N19)  
GR113-101  |     Other disorders of kidney (N25,N27)  
GR113-102  |  #Infections of kidney (N10-N12,N13.6,N15.1)  
GR113-103  |  #Hyperplasia of prostate (N40)  
GR113-104  |  #Inflammatory diseases of female pelvic organs (N70-N76)  
GR113-105  |  #Pregnancy, childbirth and the puerperium (O00-O99)  
GR113-106  |     Pregnancy with abortive outcome (O00-O07)  
GR113-107  |     Other complications of pregnancy, childbirth and the puerperium (O10-O99)  
GR113-108  |  #Certain conditions originating in the perinatal period (P00-P96)  
GR113-109  |  #Congenital malformations, deformations and chromosomal abnormalities (Q00-Q99)  
GR113-110  |  Symptoms, signs and abnormal clinical and laboratory findings, not elsewhere classified (R00-R99)  
GR113-111  |  All other diseases (Residual)   
GR113-112  |  #Accidents (unintentional injuries) (V01-X59,Y85-Y86)  
GR113-113  |     Transport accidents (V01-V99,Y85)  
GR113-114  |        Motor vehicle accidents (V02-V04...)  
GR113-115  |        Other land transport accidents (V01...)  
GR113-116  |        Water, air and space, and other and unspecified transport accidents and their sequelae (V90-V99,Y85)  
GR113-117  |     Nontransport accidents (W00-X59,Y86)  
GR113-118  |        Falls (W00-W19)  
GR113-119  |        Accidental discharge of firearms (W32-W34)  
GR113-120  |        Accidental drowning and submersion (W65-W74)  
GR113-121  |        Accidental exposure to smoke, fire and flames (X00-X09)  
GR113-122  |        Accidental poisoning and exposure to noxious substances (X40-X49)  
GR113-123  |        Other and unspecified nontransport accidents and their sequelae (W20-W31,W35-W64,W75-W99,X10-X39,X50-X59,Y86)  
GR113-124  |  #Intentional self-harm (suicide) (*U03,X60-X84,Y87.0)  
GR113-125  |     Intentional self-harm (suicide) by discharge of firearms (X72-X74)  
GR113-126  |     Intentional self-harm (suicide) by other and unspecified means and their sequelae (*U03,X60-X71,X75-X84,Y87.0)  
GR113-127  |  #Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)  
GR113-128  |     Assault (homicide) by discharge of firearms (*U01.4,X93-X95)  
GR113-129  |     Assault (homicide) by other and unspecified means and their sequelae (*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1)  
GR113-130  |  #Legal intervention (Y35,Y89.0)  
GR113-131  |  Events of undetermined intent (Y10-Y34,Y87.2,Y89.9)  
GR113-132  |     Discharge of firearms, undetermined intent (Y22-Y24)  
GR113-133  |     Other and unspecified events of undetermined intent and their sequelae (Y10-Y21,Y25-Y34,Y87.2,Y89.9)  
GR113-134  |  #Operations of war and their sequelae (Y36,Y89.1)  
GR113-135  |  #Complications of medical and surgical care (Y40-Y84,Y88)  
GR113-136  |  #Enterocolitis due to Clostridium difficile (A04.7)" %>% 
  read_lines() %>% 
  strsplit("|", fixed = TRUE) %>% 
  do.call(rbind, .) %>% 
  as_tibble() %>% 
  mutate_all(stringr::str_trim) %>% 
  mutate(V2 = tolower(V2)) %>% 
  filter(V1 %in% c(
    "GR113-112", # Accidents (unintentional injuries) (V01-X59,Y85-Y86)  
    "GR113-113", # Transport accidents (V01-V99,Y85)  
    "GR113-117", # Nontransport accidents (W00-X59,Y86)
    "GR113-127"  # Assault (homicide) (*U01-*U02,X85-Y09,Y87.1)
  )) %>% 
  mutate(V3 = case_when(
    V1 == "GR113-112" ~ "accidents",
    V1 == "GR113-113" ~ "transport_acc",
    V1 == "GR113-117" ~ "non-trans_acc",
    V1 == "GR113-127" ~ "assualt"
  )) %>% 
  print() %>% 
  write_csv("dta/03-icd10-113-cause-list.csv")
